Use a field name instead of a block to determine the category.
This is bacause when the `category` is updated, we need to decrease the original
category counter and increase the current counter. But if we are using
block to determine the category, it is hard to find out the original one.
So use a field name instead. (field original valud is easy to get)